adding Client.cpp
