+++ srcs/Client.cpp
@@ -1,51 +1,40 @@
-// Client side C/C++ program to demonstrate Socket
-// programming
-#include <arpa/inet.h>
-#include <stdio.h>
-#include <string.h>
-#include <sys/socket.h>
-#include <unistd.h>
-#include <string>
-#define PORT 8080
-  
-int main(int argc, char const* argv[])
-{
-    int status, valread, client_fd;
-    struct sockaddr_in serv_addr;
-    std::string hello = "Hello from client";
-    char buffer[1024] = { 0 };
-    if ((client_fd = socket(AF_INET, SOCK_STREAM, 0)) < 0) {
-        printf("\n Socket creation error \n");
-        return -1;
-    }
-  
-    serv_addr.sin_family = AF_INET;
-    serv_addr.sin_port = htons(PORT);
-  
-    // Convert IPv4 and IPv6 addresses from text to binary
-    // form
-    if (inet_pton(AF_INET, "127.0.0.1", &serv_addr.sin_addr)
-        <= 0) {
-        printf(
-            "\nInvalid address/ Address not supported \n");
-        return -1;
-    }
-  
-    if ((status
-         = connect(client_fd, (struct sockaddr*)&serv_addr,
-                   sizeof(serv_addr)))
-        < 0) {
-        printf("\nConnection Failed \n");
-        return -1;
-    }
-    send(client_fd, hello.c_str(), hello.size(), 0);
-    printf("Hello message sent\n");
-    valread = read(client_fd, buffer, 1024);
-    printf("%s\n", buffer);
-  
-    // closing the connected socket
-    close(client_fd);
-    return 0;
+#include "../includes/Client.hpp"
+
+Client::Client(
+    std::string nickname,
+    std::string username,
+    std::string hotsname,
+    std::string mode 
+){
+    this->nickname = nickname;
+    this->username = username;
+    this->hotsname = hotsname;
+    this->mode = mode;
 }
 
-// gcc srcs/Client.cpp -o client -lstdc++
+Client::~Client(){
+    //this->channels->clear();
+}
+
+std::string Client::get_nickname() const{
+    return this->nickname;
+}
+
+std::string Client::get_hostname() const{
+    return this->hotsname;
+}
+std::string Client::get_mode() const{
+    return this->mode;
+}
+
+//std::set<std::string> get_channels;//to write
+
+
+void Client::set_nickname(std::string nickname){
+    this->nickname = nickname;
+}
+void Client::set_mode(std::string mode){
+    this->mode = mode;
+}
+
+//void add_channel(std::string channel);//to write
